refactor: Move resolver too deep exception to its own submodule
