+++ src/pip/_internal/resolution/resolvelib/resolver.py
@@ -12,7 +12,7 @@ from pip._vendor.resolvelib import Resolver as RLResolver
 from pip._vendor.resolvelib.structs import DirectedGraph
 
 from pip._internal.cache import WheelCache
-from pip._internal.exceptions import ResolutionTooDeepError
+from pip._internal.exceptions.resolution import ResolutionTooDeepError
 from pip._internal.index.package_finder import PackageFinder
 from pip._internal.operations.prepare import RequirementPreparer
 from pip._internal.req.constructors import install_req_extend_extras
